# docs/swagger.yaml
definitions:
  admin.LoginRequest:
    properties:
      passwd:
        type: string
      user_name:
        type: string
    required:
    - passwd
    - user_name
    type: object
  admin.LoginResp:
    properties:
      token:
        type: string
    type: object
  dto.CreateOrderReq:
    properties:
      amount:
        type: integer
      channel:
        description: 充值渠道 alipay 支付宝
        type: string
      subject:
        description: 支付商品 vip 会员 amount 充值金额 chapter 购买章节
        type: string
      subject_id:
        type: integer
    type: object
  dto.CreateOrderRsp:
    properties:
      order_id:
        type: string
      qrcode:
        items:
          type: integer
        type: array
    type: object
...
  dto.Resp:
    properties:
      cate_id:
        type: integer
      code:
        type: integer
      cur_page:
        type: integer
      data:
        type: object
      msg:
        type: string
      total:
        type: integer
    type: object
...
info:
  contact:
    email: scg130@163.com
    name: scg130
  description: micro
  title: micro
  version: "1.0"
paths:
  /admin/login:
    post:
      description: 登录
      parameters:
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/admin.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  $ref: '#/definitions/admin.LoginResp'
              type: object
      summary: 登录
      tags:
      - 用户中心
  /charge/create:
    post:
      description: 创建订单
      parameters:
      - description: body参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOrderReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Resp'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateOrderRsp'
              type: object
      summary: 创建订单
      tags:
      - 订单
...
schemes:
- http
swagger: "2.0"